homework2/requirements.txt
allure-pytest==2.8.32
allure-python-commons==2.8.32
attrs==21.2.0
execnet==1.9.0
importlib-metadata==4.8.1
iniconfig==1.1.1
packaging==21.0
pluggy==1.0.0
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==2.4.0
selenium==4.0.0a5
toml==0.10.2
typing-extensions==3.10.0.2
webdriver-manager==3.3.0
zipp==3.6.0

colorama==0.4.4
pip==19.2.3
configparser==5.0.2
idna==3.3
atomicwrites==1.4.0
certifi==2021.10.8
six==1.16.0
urllib3==1.26.7
setuptools==41.2.0
requests==2.26.0